+++ src/Exception.php
@@ -1,0 +1,23 @@
+<?php
+
+/**
+ * This file is part of sata/Cache
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Slick\Cache;
+
+use Psr\Cache\CacheException;
+
+/**
+ * Exception
+ *
+ * @package Slick\Cache
+ * @author  Filipe Silva <filipe.silva@sata.pt>
+ */
+interface Exception extends CacheException
+{
+
+}